+++ dds_cli/account_manager.py
@@ -36,12 +36,19 @@ class AccountManager(dds_cli.base.DDSBaseClass):
     """Admin class for adding users, etc."""
 
     def __init__(
-        self, username: str, authenticate: bool = True, method: str = "add", no_prompt: bool = False
+        self,
+        authenticate: bool = True,
+        method: str = "add",
+        no_prompt: bool = False,
+        token_path: str = None,
     ):
         """Initialize, incl. user authentication."""
         # Initiate DDSBaseClass to authenticate user
         super().__init__(
-            username=username, authenticate=authenticate, method=method, no_prompt=no_prompt
+            authenticate=authenticate,
+            method=method,
+            no_prompt=no_prompt,
+            token_path=token_path,
         )
 
         # Only methods "add", "delete" and "revoke" can use the AccountManager class
@@ -112,7 +119,7 @@ class AccountManager(dds_cli.base.DDSBaseClass):
             LOG.info(message)
 
     def delete_own_account(self):
-        """Delete users from the system"""
+        """Delete users from the system."""
         # Perform request to API
 
         try:
@@ -125,7 +132,7 @@ class AccountManager(dds_cli.base.DDSBaseClass):
             # Get response
             response_json = response.json()
             message = response_json["message"]
-            dds_cli.auth.Auth.logout(self)
+            dds_cli.auth.Auth.logout()
 
         except requests.exceptions.RequestException as err:
             raise dds_cli.exceptions.ApiRequestError(message=str(err))
@@ -136,13 +143,13 @@ class AccountManager(dds_cli.base.DDSBaseClass):
         if not response.ok:
             if response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR:
                 raise dds_cli.exceptions.ApiResponseError(message)
-            else:
-                raise dds_cli.exceptions.DDSCLIException(message)
-        else:
-            LOG.info(message)
+
+            raise dds_cli.exceptions.DDSCLIException(message)
+
+        LOG.info(message)
 
     def revoke_project_access(self, project, email):
-        """Revoke a user's access to a project"""
+        """Revoke a user's access to a project."""
         json = {"email": email}
         try:
             response = requests.post(
@@ -175,7 +182,7 @@ class AccountManager(dds_cli.base.DDSBaseClass):
         )
 
     def get_user_info(self):
-        """Get a users info"""
+        """Get a users info."""
         try:
             response = requests.get(
                 dds_cli.DDSEndpoint.DISPLAY_USER_INFO,