+++ dds_cli/data_getter.py
@@ -40,7 +40,6 @@ class DataGetter(base.DDSBaseClass):
 
     def __init__(
         self,
-        username: str,
         project: str = None,
         break_on_fail: bool = False,
         get_all: bool = False,
@@ -51,15 +50,16 @@ class DataGetter(base.DDSBaseClass):
         verify_checksum: bool = False,
         method: str = "get",
         no_prompt: bool = False,
+        token_path: str = None,
     ):
         """Handle actions regarding downloading data."""
         # Initiate DDSBaseClass to authenticate user
         super().__init__(
-            username=username,
             project=project,
             dds_directory=destination,
             method=method,
             no_prompt=no_prompt,
+            token_path=token_path,
         )
 
         # Initiate DataGetter specific attributes