+++ dds_cli/data_putter.py
@@ -43,7 +43,6 @@ LOG = logging.getLogger(__name__)
 
 
 def put(
-    username,
     project,
     source,
     source_path_file,
@@ -52,11 +51,11 @@ def put(
     num_threads,
     silent,
     no_prompt,
+    token_path,
 ):
     """Handle upload of data."""
     # Initialize delivery - check user access etc
     with DataPutter(
-        username=username,
         project=project,
         source=source,
         source_path_file=source_path_file,
@@ -64,6 +63,7 @@ def put(
         overwrite=overwrite,
         silent=silent,
         no_prompt=no_prompt,
+        token_path=token_path,
     ) as putter:
 
         # Progress object to keep track of progress tasks
@@ -171,7 +171,6 @@ class DataPutter(base.DDSBaseClass):
 
     def __init__(
         self,
-        username: str,
         project: str = None,
         break_on_fail: bool = False,
         overwrite: bool = False,
@@ -180,10 +179,16 @@ class DataPutter(base.DDSBaseClass):
         silent: bool = False,
         method: str = "put",
         no_prompt: bool = False,
+        token_path: str = None,
     ):
         """Handle actions regarding upload of data."""
         # Initiate DDSBaseClass to authenticate user
-        super().__init__(username=username, project=project, method=method, no_prompt=no_prompt)
+        super().__init__(
+            project=project,
+            method=method,
+            no_prompt=no_prompt,
+            token_path=token_path,
+        )
 
         # Initiate DataPutter specific attributes
         self.break_on_fail = break_on_fail